requirements.txt
streamlit==1.32.0
cx_Oracle==8.3.0
groq==0.5.0
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.2.0
plotly==5.18.0